backend/requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
pytest
pydantic>=2.0
requests
scikit-learn
numpy
openai
